django-backend/pyproject.toml
[tool.poetry]
name = "be-news-app"
version = "0.1.0"
description = ""
authors = ["Ahmed Osama <h.ahmedosama@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
djangorestframework = "^3.14.0"
newsapi-python = "^0.2.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
